removing my signature (#130)
This seems to be an empty display of morality, as each signatory can interpret the Oath however suits them.

An oath that avoids taking principled stands so that it can be all things to all people isn't much of an oath.

If at least the oath hosted a dialog about what the oath means, that discussion would be very valuable in moving tech culture in a positive, morals over money direction. But issues raised get silenced or dismissed (see Issue #120). 

"It is difficult to get a man to understand something when his salary depends upon his not understanding it.” – Upton Sinclair

This is a big deal for me because I've dedicated my life toward social justice and fairness issues. The tech world, sadly, is too focused on money. 

Good luck.